third_party/blink/web_tests/platform/generic/external/wpt/css/css-contain/container-queries/custom-property-style-queries-expected.txt

This is a testharness.js-based test.
FAIL style(--inner: true) assert_equals: expected "true" but got ""
FAIL style(--inner:true) assert_equals: expected "true" but got ""
FAIL style(--inner:true ) assert_equals: expected "true" but got ""
FAIL style(--inner: true ) assert_equals: expected "true" but got ""
PASS style(--inner-no-space: true)
PASS style(--inner-no-space:true)
FAIL style(--inner-no-space:true ) assert_equals: expected "true" but got ""
FAIL style(--inner-no-space: true ) assert_equals: expected "true" but got ""
FAIL style(--inner-space-after: true) assert_equals: expected "true" but got ""
FAIL style(--inner-space-after:true) assert_equals: expected "true" but got ""
PASS style(--inner-space-after:true )
PASS style(--inner-space-after: true )
PASS style(--outer: true)
PASS style(--outer:true)
PASS style(--outer:true )
PASS style(--outer: true )
PASS style(--outer-no-space: true)
PASS style(--outer-no-space:true)
PASS style(--outer-no-space:true )
PASS style(--outer-no-space: true )
PASS style(--outer-space-after: true)
PASS style(--outer-space-after:true)
PASS style(--outer-space-after:true )
PASS style(--outer-space-after: true )
PASS outer style(--inner: true)
PASS outer style(--inner:true)
PASS outer style(--inner:true )
PASS outer style(--inner: true )
PASS outer style(--inner-no-space: true)
PASS outer style(--inner-no-space:true)
PASS outer style(--inner-no-space:true )
PASS outer style(--inner-no-space: true )
PASS outer style(--inner-space-after: true)
PASS outer style(--inner-space-after:true)
PASS outer style(--inner-space-after:true )
PASS outer style(--inner-space-after: true )
FAIL outer style(--outer: true) assert_equals: expected "true" but got ""
FAIL outer style(--outer:true) assert_equals: expected "true" but got ""
FAIL outer style(--outer:true ) assert_equals: expected "true" but got ""
FAIL outer style(--outer: true ) assert_equals: expected "true" but got ""
PASS outer style(--outer-no-space: true)
PASS outer style(--outer-no-space:true)
FAIL outer style(--outer-no-space:true ) assert_equals: expected "true" but got ""
FAIL outer style(--outer-no-space: true ) assert_equals: expected "true" but got ""
FAIL outer style(--outer-space-after: true) assert_equals: expected "true" but got ""
FAIL outer style(--outer-space-after:true) assert_equals: expected "true" but got ""
PASS outer style(--outer-space-after:true )
PASS outer style(--outer-space-after: true )
Harness: the test ran to completion.
